Refactor WIP
