+++ wxm-ios/Toolkit/Toolkit/Logger/LoggerTypes.swift
@@ -11,14 +11,3 @@ public protocol NetworkError: Error {
     var code: Int { get }
     var userInfo: [String: Any]? { get }
 }
-
-protocol LoggerImplementation {
-	func logNetworkError(_ networkError: NetworkError)
-	func logError(_ nsError: NSError)
-	func trackScreen(_ screen: Screen, parameters: [Parameter: ParameterValue]?)
-	func trackEvent(_ event: Event, parameters: [Parameter: ParameterValue]?)
-	func setUserId(_ userId: String?)
-	func setUserProperty(key: Parameter, value: ParameterValue)
-	func setDefaultParameter(key: Parameter, value: ParameterValue)
-	func setAnalyticsCollectionEnabled(_ enabled: Bool)
-}
